Persistence Fixes (#115)
* isPersistenceEnabled added
* Fixes handling of Private windows for Safari and Firefox
* Fixes handling of downgrading to older WebSDK version

+++ src/db-manager.js
@@ -18,7 +18,7 @@ const SyncEvent = require('./sync-event');
 const Constants = require('./const');
 const Util = require('./client-utils');
 
-const DB_VERSION = 20;
+const DB_VERSION = 2;
 const MAX_SAFE_INTEGER = 9007199254740991;
 const SYNC_NEW = Constants.SYNC_STATE.NEW;
 
@@ -67,6 +67,7 @@ class DbManager extends Root {
     super(options);
 
     // If no indexedDB, treat everything as disabled.
+    /* istanbul ignore next */
     if (!window.indexedDB) {
       options.tables = {};
     } else {
@@ -104,18 +105,22 @@ class DbManager extends Root {
     TABLES.forEach((tableDef) => {
       this['_permission_' + tableDef.name] = Boolean(options.tables[tableDef.name]);
     });
-    this._open();
+    this._open(false);
   }
 
+  _getDbName() {
+    return 'LayerWebSDK_' + this.client.appId + '_' + this.client.user.userId.replace(/[^a-zA-Z0-9]/g, '');
+  }
 
   /**
    * Open the Database Connection.
    *
    * This is only called by the constructor.
    * @method _open
+   * @param {Boolean} retry
    * @private
    */
-  _open() {
+  _open(retry) {
     if (this.db) {
       this.db.close();
       delete this.db;
@@ -131,13 +136,21 @@ class DbManager extends Root {
 
     // Open the database
     const client = this.client;
-    const request = window.indexedDB.open('LayerWebSDK_' + client.appId + '_' + client.user.userId.replace(/[^a-zA-Z0-9]/g, ''), DB_VERSION);
+    const request = window.indexedDB.open(this._getDbName(), DB_VERSION);
 
     try {
       request.onerror = (evt) => {
-        this._isOpenError = true;
-        logger.error('Database Unable to Open: ', evt.target.error);
-        this.trigger('error', { error: evt });
+        if (!retry) {
+          this.deleteTables(() => this._open(true));
+        }
+
+        // Triggered by Firefox private browsing window
+        /* istanbul ignore next */
+        else {
+          this._isOpenError = true;
+          logger.warn('Database Unable to Open (common cause: private browsing window)', evt.target.error);
+          this.trigger('error', { error: evt });
+        }
       };
 
       request.onupgradeneeded = (evt) => this._onUpgradeNeeded(evt);
@@ -151,9 +164,7 @@ class DbManager extends Root {
           this.isOpen = false;
         };
 
-        this.db.error = err => {
-          logger.error('db-manager Error: ', err);
-        };
+        this.db.onerror = err => logger.error('db-manager Error: ', err);
       };
     }
 
@@ -191,19 +202,21 @@ class DbManager extends Root {
   /* istanbul ignore next */
   _onUpgradeNeeded(event) {
     const db = event.target.result;
+    let isComplete = false;
 
-    let completeCount = 0;
-    function onComplete() {
-      completeCount++;
-      if (completeCount === TABLES.length) {
+    // This appears to only get called once; its presumed this is because we're creating but not using a lot of transactions.
+    var onComplete = (evt) => {
+      if (!isComplete) {
+        this.db = db;
         this.isOpen = true;
         this.trigger('open');
       }
-    }
+    };
 
+    const currentTables = Array.prototype.slice.call(db.objectStoreNames);
     TABLES.forEach((tableDef) => {
       try {
-        db.deleteObjectStore(tableDef.name);
+        if (currentTables.indexOf(tableDef.name) !== -1) db.deleteObjectStore(tableDef.name);
       } catch (e) {
         // Noop
       }
@@ -1080,16 +1093,14 @@ class DbManager extends Root {
    * @param {Function} [calllback]
    */
   deleteTables(callback = function() {}) {
-    this.onOpen(() => {
-      try {
-        var request = window.indexedDB.deleteDatabase(this.db.name);
-        request.onsuccess = callback;
-        delete this.db;
-       } catch (e) {
-        logger.error('Failed to delete database', e);
-        if (callback) callback(e);
-      }
-    });
+    try {
+      var request = window.indexedDB.deleteDatabase(this._getDbName());
+      request.onsuccess = request.onerror = callback;
+      delete this.db;
+    } catch (e) {
+      logger.error('Failed to delete database', e);
+      if (callback) callback(e);
+    }
   }
 }
